@@ -1,5 +1,6 @@
 require 'rubygems'
 require 'spork'
+ENV["RAILS_ENV"] ||= 'test'
 
 Spork.prefork do
   # Loading more in this block will cause your tests to run faster. However,
@@ -7,9 +8,9 @@ Spork.prefork do
   # need to restart spork for it take effect.
   # This file is copied to spec/ when you run 'rails generate rspec:install'
 
-  ENV["RAILS_ENV"] ||= 'test'
   require File.expand_path("../../config/environment", __FILE__)
   require 'rspec/rails'
+  require 'capybara/rspec'
 
   # Requires supporting ruby files with custom matchers and macros, etc,
   # in spec/support/ and its subdirectories.
@@ -23,8 +24,10 @@ Spork.prefork do
     config.use_transactional_fixtures = true
 
     config.treat_symbols_as_metadata_keys_with_true_values = true
-    config.filter_run :focus => true
+    config.filter_run focus: true
     config.run_all_when_everything_filtered = true
+
+    config.include Capybara::DSL
 
     # For original DatabaseCleaner setting watch github repo at https://github.com/bmabey/database_cleaner
     DatabaseCleaner.strategy = :truncation
