@@ -9,6 +9,6 @@ ActiveSupport.on_load(:action_controller) do
 end
 
 # Disable root element in JSON by default.
-ActiveSupport.on_load(:active_record) do
-  self.include_root_in_json = false
-end
+# ActiveSupport.on_load(:active_record) do
+#   self.include_root_in_json = false
+# end
